refactor: name & struct
