+++ web/app/components/apply/rule_yara_preview_card.tsx
