+++ web/app/components/apply/rule_yara_select_card.tsx
